requirements.txt
amqp==5.1.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.4
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.3.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
